Playing Around/requirements.txt
msal>=1.28.0
python-dotenv>=1.0.0
requests>=2.32.0
